Encapsulate forge helpers and split test suites
- Keep provider-specific command construction, parsing, and URL generation within their owning implementations while centralizing command timeout execution.
- Split large forge unit suites into behavior-focused modules with explicit imports and shared fixtures.
- Expand coverage for command execution failures, remote parsing, review anchors, metadata synchronization, and malformed responses.
