+++ crates/ag-forge/src/github_test.rs
@@ -1,0 +1,10 @@
+#[path = "github_test/comment_test.rs"]
+mod comment_test;
+#[path = "github_test/metadata_test.rs"]
+mod metadata_test;
+#[path = "github_test/parse_test.rs"]
+mod parse_test;
+#[path = "github_test/request_test.rs"]
+mod request_test;
+#[path = "github_test/support_test.rs"]
+mod support_test;
