+++ crates/ag-forge/src/gitlab_test.rs
@@ -1,0 +1,10 @@
+#[path = "gitlab_test/comment_test.rs"]
+mod comment_test;
+#[path = "gitlab_test/metadata_test.rs"]
+mod metadata_test;
+#[path = "gitlab_test/parse_test.rs"]
+mod parse_test;
+#[path = "gitlab_test/request_test.rs"]
+mod request_test;
+#[path = "gitlab_test/support_test.rs"]
+mod support_test;
